// src/core/models/planet.ts
interface IPlanet {
  id: number;
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: number[];
  films: number[];
  created: string;
  edited: string;
  url: string;
}

export default IPlanet;